backend/voice/requirement.txt
fastapi               
uvicorn                
faster-whisper       
requests             
python-multipart      
python-dotenv        